% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/PCPM-class.R
\name{PCPMr}
\alias{PCPMr}
\title{PCPMr R6 Class}
\description{
This class wraps the PCPM C++ class using R6.
}
\examples{
# example code

\donttest{
truemodel<-sim_matrix(n_specifics=3, items_per_specific=6,loadings_s = c(.7,.7),lac_spe=0.2, cpf_spe=2,specifics_rho = 0.3, seed=123)
dat <- sim_data( N = 100,lam = truemodel[["A"]],phi = truemodel[["Phi"]],ecm =truemodel[["ecm"]])
J<-nrow(dat$lam);K<-ncol(dat$lam)
Q <- matrix(-1, nrow = J, ncol = K)
Q[dat$lam!=0]=-2
Y<-dat$dat
a<-PCPMr$new()
a$pcfa(Y, Q, LD=F,iter=100, burn=0,std=T, cor_fac=1, update = 100)
out<-a$pcpmsummary(sig=F)
}
}
\section{Public fields}{
\if{html}{\out{<div class="r6-fields">}}
\describe{
\item{\code{N}}{Number of observations.}

\item{\code{J}}{Number of items.}

\item{\code{K}}{Number of latent factors.}

\item{\code{P}}{Number of predictors.}

\item{\code{Q}}{Loading pattern matrix.}

\item{\code{Qb}}{Regression pattern matrix.}

\item{\code{LD}}{Logical indicating local dependence.}

\item{\code{DIF}}{Logical indicating differential item functioning.}

\item{\code{time}}{Computation time.}

\item{\code{MODEL}}{Model type identifier.}

\item{\code{FacSco}}{Stored factor scores.}
}
\if{html}{\out{</div>}}
}
\section{Methods}{
\subsection{Public methods}{
\itemize{
\item \href{#method-PCPMr-new}{\code{PCPMr$new()}}
\item \href{#method-PCPMr-pcfa}{\code{PCPMr$pcfa()}}
\item \href{#method-PCPMr-getVariable}{\code{PCPMr$getVariable()}}
\item \href{#method-PCPMr-pcpmsummary}{\code{PCPMr$pcpmsummary()}}
\item \href{#method-PCPMr-clone}{\code{PCPMr$clone()}}
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-PCPMr-new"></a>}}
\if{latex}{\out{\hypertarget{method-PCPMr-new}{}}}
\subsection{Method \code{new()}}{
Initializes a new PCPM object.
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{PCPMr$new(
  priorF = c(0.1),
  priorA = c(1, 0.1, 1, 0.1, 0, 0.1),
  priorB = c(1, 0.1, 1, 0.1, 0, 0.1)
)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{priorF}}{A numeric vector for priorF (default: c(0.1, 1.0, 1.0)). w0}

\item{\code{priorA}}{A numeric vector for priorA (default: c(1.0, 1.0, 1.0, 1.0, 0.0, 1.0)).s0;r0;lam_a;lam_b;m0;c0;}

\item{\code{priorB}}{A numeric vector for priorB (default: c(1.0, 1.0, 1.0, 1.0, 0.0, 1.0)).s0B;r0B;lam_aB;lam_bB;m0B;c0B;}
}
\if{html}{\out{</div>}}
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-PCPMr-pcfa"></a>}}
\if{latex}{\out{\hypertarget{method-PCPMr-pcfa}{}}}
\subsection{Method \code{pcfa()}}{
Performs PCFA on the provided data.
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{PCPMr$pcfa(
  Y,
  Q,
  cati = F,
  cati_v = -1,
  reglo = "lasso",
  LD = F,
  regphi = "lasso",
  Kg = 0,
  reg,
  regpsx = "lasso",
  iter = 1000,
  burn = 0,
  std = T,
  cor_fac = 1,
  sign_check = F,
  update = 1000
)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{Y}}{A numeric matrix.}

\item{\code{Q}}{An integer matrix.}

\item{\code{cati}}{A logical value, wheter item is categorical.}

\item{\code{cati_v}}{A vector or -1. which item is categorical, -1 means all.}

\item{\code{reglo}}{A string. which prior for loading. 'lasso','horse','ssp'}

\item{\code{LD}}{A logical value.}

\item{\code{regphi}}{A string. which prior for factor correlation. 'none', lasso','horse','ssp','parlasso','parhorse','parssp'}

\item{\code{Kg}}{A numeric. how many general factor, if none, 0}

\item{\code{reg}}{An integer matrix. set for par phi, not used here. default as matrix full with 1.}

\item{\code{regpsx}}{A string. which prior for loading. 'lasso','horse','ssp'}

\item{\code{iter}}{An integer value for the number of iterations.}

\item{\code{burn}}{An integer value for the burn-in period.}

\item{\code{std}}{A logical value indicating whether to standardize.}

\item{\code{cor_fac}}{A numeric vector for correlation factors.}

\item{\code{sign_check}}{A logical value.}

\item{\code{update}}{An integer value for the update frequency (default: 1000).}
}
\if{html}{\out{</div>}}
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-PCPMr-getVariable"></a>}}
\if{latex}{\out{\hypertarget{method-PCPMr-getVariable}{}}}
\subsection{Method \code{getVariable()}}{
get posterial summary
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{PCPMr$getVariable(varName)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{varName}}{aaa.}
}
\if{html}{\out{</div>}}
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-PCPMr-pcpmsummary"></a>}}
\if{latex}{\out{\hypertarget{method-PCPMr-pcpmsummary}{}}}
\subsection{Method \code{pcpmsummary()}}{
Performs pcpmsummary.
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{PCPMr$pcpmsummary(
  med = FALSE,
  start = 0,
  end = -1,
  SL = 0.05,
  sig = T,
  detail = F,
  SignSwitch = F,
  reorder = F
)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{med}}{aaa.}

\item{\code{start}}{aaa.}

\item{\code{end}}{aaaa}

\item{\code{SL}}{aaa}

\item{\code{sig}}{aaa}

\item{\code{detail}}{aaa}

\item{\code{SignSwitch}}{aaa}

\item{\code{reorder}}{aaa}
}
\if{html}{\out{</div>}}
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-PCPMr-clone"></a>}}
\if{latex}{\out{\hypertarget{method-PCPMr-clone}{}}}
\subsection{Method \code{clone()}}{
The objects of this class are cloneable with this method.
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{PCPMr$clone(deep = FALSE)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{deep}}{Whether to make a deep clone.}
}
\if{html}{\out{</div>}}
}
}
}
